refactor: 스텝별 폴더 생성 및 이동
