+++ app/user-web-prototype/src/components/presentational/model-learning/data-navigation/DataNavigationContentTable.jsx
@@ -5,7 +5,7 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import {Checkbox, Paper, Switch} from "@mui/material";
 import TableBody from "@mui/material/TableBody";
-import {StyledTable, StyledTableCell} from "./StyledTableComponents";
+import {StyledTable, StyledTableCell} from "../StyledTableComponents";
 
 const AVERAGE_SPEED = "평균속도";
